+++ GE/worldStructFunctions.go
@@ -16,8 +16,8 @@ func (p *WorldStructure) ToBytes() ([]byte, error) {
 	
 	mdxBs := cmp.Int64ToBytes(int64(p.middleX))
 	mdyBs := cmp.Int64ToBytes(int64(p.middleY))
-	minBs := cmp.Int16ToBytes(p.minLight)
-	maxBs := cmp.Int16ToBytes(p.maxLight)
+	minBs := cmp.Int16ToBytes(int16(p.minLight))
+	maxBs := cmp.Int16ToBytes(int16(p.maxLight))
 	delBs := cmp.Float64ToBytes(p.deltaB)
 	return cmp.CompressAll(changing, mdxBs, mdyBs, minBs, maxBs, delBs), nil
 }
@@ -36,7 +36,7 @@ func GetWorldStructureFromBytes(X,Y,W,H float64, data []byte, tile_path, struct_
 	p.BytesToObjects(bs[6])
 	p.BytesToLights(bs[7])
 	p.SetMiddle(int(cmp.BytesToInt64(bs[0])), int(cmp.BytesToInt64(bs[1])), true)
-	p.SetLightStats(cmp.BytesToInt16(bs[2]), cmp.BytesToInt16(bs[3]), cmp.BytesToFloat64(bs[4]))
+	p.SetLightStats(int64(cmp.BytesToInt16(bs[2])), int64(cmp.BytesToInt16(bs[3])), cmp.BytesToFloat64(bs[4]))
 	return p, nil
 }
